define LimitsReachedError in settings as dotted path to import
